+++ api/database/prismaClient.js
@@ -1,0 +1,4 @@
+const { PrismaClient } = require('@prisma/client');
+const prisma = new PrismaClient();
+
+module.exports = prisma;
